Deleted UserPage and added LocationPage table (#101)
* Deleted UserPage and added LocationPage table

* Ran prettier & lint

---------
