+++ client/src/Pages/index.jsx
@@ -4,6 +4,5 @@ import Clients from './Clients';
 import Inventory from './Inventory';
 import Locations from './Locations';
 import Donors from './Donors';
-import Users from './Users';
 
-export { Appointments, Tickets, Clients, Inventory, Locations, Donors, Users };
+export { Appointments, Tickets, Clients, Inventory, Locations, Donors };
